autonomous_agent/requirements.txt
playwright
pillow==11.1.0
openai
python-dotenv
pyyaml
